Fixed search bar
The search bar in the home page should now be functional. Don't mind the code tho i couldnt be assed to try and fix the function from the sql server so i just coded it locally into the python file. Also fixed the css and html related to the search so it doesnt look too bad